pave the way for a new decompression system
the fundamental problem that is posed by our current streambase is as follows: for the compression, the streambase has to read from 1 file into the buffer, compress, and write to 2 files. For the decompression however, the streambase has to read from 2 files, decompress, and write to the buffer. The current way things are handled, which is, take a filepath, concatenate .meta and .data respectively and do your stuff, which is ideal for the compression, but not so much for the decompression, as I would like to open 2 separate streams for meta and data respectively. This new system should accomodate for that.

other things that changed:

- taken the bkyin and decomp from bkystreambase
- nice lil status messages for lzmaio which are now commented out
- NEVER PRINT BUFFERS TO CONSOLE